chg: reorganized hookspecs
