+++ src/quilla/hookspecs/__init__.py
@@ -1,0 +1,3 @@
+import pluggy
+
+hookspec = pluggy.HookspecMarker('quilla')
